requirements.txt
asyncpraw==7.4.0
asyncprawcore==2.3.0
async-timeout==3.0.1
asyncio-extras==1.3.2
cs50
discord==1.7.3
discord-ext-bot==1.0.1
discord.py==1.7.3
python-dotenv==0.19.1
discord-pretty-help
